third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/the-innerhtml-property/innerhtml-and-xml-namespaces-expected.txt

This is a testharness.js-based test.
[FAIL] default namespace applies to children of non-namespaced element
  assert_equals: expected (object) null but got (string) "http://www.w3.org/2000/svg"
[FAIL] default namespace applies to children of namespaced element with new default namespace
  assert_equals: expected "https://example.org/ns" but got "http://www.w3.org/2000/svg"
[FAIL] unsetting default namespace works inside parse of fragment (on element)
  assert_equals: expected (object) null but got (string) "http://www.w3.org/2000/svg"
[FAIL] unsetting default namespace works inside parse of fragment (on child of element)
  assert_equals: expected (object) null but got (string) "http://www.w3.org/2000/svg"
Harness: the test ran to completion.
